Add AST generation for functions.
This commit makes it possible to generate the AST for a function by
passing the function's name to `to_ast_()`. For example, to generate the
AST for the `lm()` function, use `to_ast_(lm)`.

The commit adds two new subclasses of ASTNode: Callable and Primitive.
The Callable class is an abstract class for representing callable
objects, which currently means functions and primitives. The Primitive
class differentiates primitives from functions so that they can be
handled separately by users.