Add AST generation for anonymous function calls.
This commit implements AST generation for calls to anonymous functions,
such as `(function(x) 3)(12)`. The Call class now has a `func` field,
which holds a Function object or a name (character), and no longer has a
`name` field. The Return class is no longer a subclass of the Call class
and no longer has a `name` field.