Overhaul parent tracking for ASTNode objects.
This commit improves the existing API for creating ASTNode objects by
changing the way parent-tracking is handled. Creating an ASTNode object
no longer require an explicitly specified parent. Instead, parent nodes
are responsible for setting the parent reference on all of their
children. This reduces the amount of boilerplate needed to create new
nodes.

Control-flow graph generation now inserts code to define iterator
variables in for-loops. As a consequence, the CFG for a program may
contain code that doesn't appear in the AST, although both
representations still have equivalent meaning.

This commit also adds a Phi class in preparation for translation to
single-static assignment form.