Fix and refactor cross-block use collection.
This commit fixes cross-block use collection to detect replacements and
function arguments as uses. Cross-block use collection is an important
step in converting a CFG to SSA form.

This commit also moves the code for cross-block use collection into a
separate function, `collect_crossblock_uses()`, so that the analysis can
be reused.