Make to_ast convert parameters for primitives.
This commit makes to_ast convert parameters for primitives to Parameter
objects, so that to_ast is consistent across primitives and functions.
This commit also changes the order of arguments to Primitive$initialize
to be more like the Function$initialize. Finally, this commit fixes a
bug in how Primitives were being initialized.