Switch to igraph-based flow graphs.
This commit is NOT stable. This commit is the first in a sequence that
replaces the previous reference-based graph data structure with igraph.
This has many benefits, but the main motivation is fast, convenient
access to edge sets.

The functionality of the CFGraph class is now split across the FlowGraph
and CFGBuilder classes. The FlowGraph class represents a control-flow
graph as an igraph and list of basic blocks. The CFGBuilder class
contains methods used by the `to_cfg` function to build CFGs, which
allows the FlowGraph class to be simpler.