Make Replacement a subclass of Assign.
This commit changes the Replacement class to be a subclass of Assign
instead of Call. The call part of the replacement is now held in the
`read` field of the Replacement object. This ensures that different SSA
names are used for variables before and after replacement, and that we
can access both easily.

As an added benefit, several methods for Replacement have been removed
since the methods for Assign can be used instead.