Fix bug in SSA construction for loops.
This commit fixes a bug in the phi-insertion stage of SSA construction.
In particular, the phi-insertion stage was not looping over all elements
of the worklist. As a result,  some phi-functions were not being placed,
notably for assignments in if-statements in for-loops. This commit also
adds a test for that case.