Change over to lightweight CFG generation.
This commit replaces `to_cfg()` with a new version that generates the
CFG without breaking the AST. The `to_ssa()` function is still not
working yet.

This also allows for a significant reduction in lines of code. Until
`to_ssa()` is working, the old old code is in the `old_` files.