Switch to pruned SSA.
This commit uses the new data flow analysis functions to generate pruned
SSA instead of minimal SSA. In pruned SSA, phi-functions are only added
for a variable when the variable will be used later in the code. This
contrasts minimal SSA, where phi-functions are always added for a
variable if that variable is used in more than one basic block.