Fix up functions to replace and insert nodes.
This commit updates Duncan's functions to replace and insert nodes so
that they work with the new class hierarchy. This commit also adds a
CodeList class for nodes that have a list of children (rather than a
fixed number of children). CodeList is a superclass for Container,
Callable, and Application.