Fix name collision for get_index, adding where_is.
This commit fixes a name collision between two different versions of the
`get_index` function. The older version finds the index of a node in its
tree, while the newer version gets the index arguments of a call to a
subset function. The older version is now called `where_is`.