Add forward data flow analysis solver.
This commit leaves parts of the package broken, to be fixed in
subsequent commits. The commit adds a function to solve forward data
flow analyses, and also changes the function to solve backward data flow
analyses to use logical vectors. The new solvers are not yet integrated
with the to_ssa function (the only function in the package that uses
them).